Rename file: LayerByLayerDiagramMaker -> StructureDiagramMaker
