+++ src/StructureDiagramMaker.js
